uk-pmrv-app-web/src/app/tasks/aer/verification-submit/compliance-ets/non-conformities/non-conformities.component.ts
import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
import { UntypedFormGroup } from '@angular/forms';

import { AER_TASK_FORM } from '@tasks/aer/core/aer-task-form.token';
import { nonConformitiesFormProvider } from '@tasks/aer/verification-submit/compliance-ets/non-conformities/non-conformities.form.provider';
import { ComplianceEtsService } from '@tasks/aer/verification-submit/compliance-ets/services/compliance-ets.service';

@Component({
  selector: 'app-non-conformities',
  templateUrl: './non-conformities.component.html',
  providers: [nonConformitiesFormProvider, ComplianceEtsService],
  changeDetection: ChangeDetectionStrategy.OnPush,
})
export class NonConformitiesComponent {
  isEditable$ = this.complianceEtsService.isEditable$;

  constructor(
    @Inject(AER_TASK_FORM) readonly form: UntypedFormGroup,
    private readonly complianceEtsService: ComplianceEtsService,
  ) {}

  onSubmit(): void {
    this.complianceEtsService.onSubmit('../summary', this.form.dirty, this.form.value);
  }
}
